refactor : Update day 1 and day 2
