+++ 2022/day-1/data.js
@@ -1,4 +1,4 @@
-const { readData } = require("../../helpers/data");
+const { readData } = require("../../helpers/input");
 
 const data = readData(__dirname).split("\n\n");
