+++ 2022/day-2/README.md
@@ -1,4 +1,4 @@
-# 🍪 Day 1 – Calorie Counting
+# ✂️ Day 2 – Rock Paper Scissors
 
 The Elves begin to set up camp on the beach. To decide whose tent gets to be closest to the snack storage, a giant Rock Paper Scissors tournament is already in progress.
